Fix bug with new line at end of block
Engine.update had a bug with how it was resolving cells in the case of
it being triggered at the start of a new line within a block.

The code that was causing the bug was actually redundant and got
fully removed.

Also updates some documentation and adds regression tests